Refactor Matchers/Extractor into their own file
